requirements.txt
python-telegram-bot
requests
urllib3
aiohttp
asyncio
python-dotenv
aiogram
telethon